Big Commit!
We can Now Assign Values to INTEGER's, and do arithmetic with those values.

IF, WHILE's, FOR's, & char's still need to finished. I am exhausted, it took me forever to figure this thing out. Like 10 hrs +

I added functions and a variable "Token* value" to the SymbolTable.h.

EXPLANATION: I had to do this because we need somewhere to store the values once we assign them. This took me forever even though it was easy lol.

So the Output basically prints out how it goes through the Interpreter.h.
We added colors to Terminal to make it easier to follow.